backend/pipeline/evaluation/requirements.txt
functions-framework==3.*
google-cloud-pubsub>=2.0.0